requirements.txt
requests==2.32.3
google-generativeai==0.8.3
langchain==0.3.13
python-dotenv==1.0.1